Adicionando logica a class Application para share de dados
