+++ app/src/main/AndroidManifest.xml
@@ -3,7 +3,7 @@
     package="com.example.android.app">
 
     <application
-        android:name=".MerchantApplication"
+        android:name=".MyApplication"
         android:allowBackup="true"
         android:icon="@drawable/ic_launcher"
         android:label="@string/app_name"
